size类型修改为size_t解决数值过大问题

+++ ais-bench_workload/tool/ais_infer/backend/base/module/ModelInfer/ModelInferenceProcessor.cpp
@@ -157,7 +157,7 @@ APP_ERROR ModelInferenceProcessor::AddOutTensors(std::vector<MemoryData>& output
 {
     bool is_dymshape = (dynamicInfo_.dynamicType == DYNAMIC_SHAPE ? true : false);
     uint64_t dymbatch_size = (dynamicInfo_.dynamicType == DYNAMIC_BATCH ? dynamicInfo_.dyBatch.batchSize : 0);
-    int realLen;
+    size_t realLen;
     for (const auto& name : outputNames) {
         auto index = modelDesc_.outnames2Index[name];
 
@@ -168,7 +168,7 @@ APP_ERROR ModelInferenceProcessor::AddOutTensors(std::vector<MemoryData>& output
             // 针对于动态shape场景 如果无法获取真实的输出shape 先填写一个一维的值 以便后续内存可以导出
             i64shape.push_back(realLen / aclDataTypeSize(static_cast<aclDataType>(modelDesc_.outTensorsDesc[index].datatype)));
         }
-        DEBUG_LOG("AddOutTensors name:%s index:%d len:%d outdescsize:%zu shapesize:%zu",
+        DEBUG_LOG("AddOutTensors name:%s index:%d len:%zu outdescsize:%zu shapesize:%zu",
             name.c_str(), index, realLen, modelDesc_.outTensorsDesc[index].size, i64shape.size());
         outputs[index].size = realLen;
         bool isBorrowed = false;
